+++ src/main/java/com/webtutsplus/ecommerce/model/Category.java
@@ -10,7 +10,7 @@ public class Category {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
 	@Column(name = "category_name")
 	private @NotBlank String categoryName;
@@ -24,10 +24,10 @@ public class Category {
 	@OneToMany(mappedBy = "category", fetch = FetchType.LAZY,
 			cascade = CascadeType.ALL)
 	Set<Product> products;
-	
+
 	public Category() {
 	}
-	
+
 	public Category(@NotBlank String categoryName, @NotBlank String description) {
 		this.categoryName = categoryName;
 		this.description = description;
@@ -76,11 +76,11 @@ public class Category {
 		this.imageUrl = imageUrl;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 }
