+++ src/main/java/com/webtutsplus/ecommerce/service/CategoryService.java
@@ -15,15 +15,15 @@ import com.webtutsplus.ecommerce.repository.Categoryrepository;
 public class CategoryService {
 
 	private final Categoryrepository categoryrepository;
-	
+
 	public CategoryService(Categoryrepository categoryrepository) {
 		this.categoryrepository = categoryrepository;
 	}
-	
+
 	public List<Category> listCategories() {
 		return categoryrepository.findAll();
 	}
-	
+
 	public void createCategory(Category category) {
 		categoryrepository.save(category);
 	}
@@ -32,11 +32,11 @@ public class CategoryService {
 		return categoryrepository.findByCategoryName(categoryName);
 	}
 
-	public Optional<Category> readCategory(Long categoryId) {
+	public Optional<Category> readCategory(Integer categoryId) {
 		return categoryrepository.findById(categoryId);
 	}
 
-	public void updateCategory(Long categoryID, Category newCategory) {
+	public void updateCategory(Integer categoryID, Category newCategory) {
 		Category category = categoryrepository.findById(categoryID).get();
 		category.setCategoryName(newCategory.getCategoryName());
 		category.setDescription(newCategory.getDescription());
