+++ src/main/java/com/webtutsplus/ecommerce/service/ProductService.java
@@ -42,14 +42,14 @@ public class ProductService {
         productRepository.save(product);
     }
 
-    public void updateProduct(long productID, ProductDto productDto, Category category) {
+    public void updateProduct(Integer productID, ProductDto productDto, Category category) {
         Product product = getProductFromDto(productDto, category);
         product.setId(productID);
         productRepository.save(product);
     }
 
 
-    public Product getProductById(Long productId) throws ProductNotExistException {
+    public Product getProductById(Integer productId) throws ProductNotExistException {
         Optional<Product> optionalProduct = productRepository.findById(productId);
         if (!optionalProduct.isPresent())
             throw new ProductNotExistException("Product id is invalid " + productId);
